aider: Added rules to ignore CMake generated files in .gitignore.
# Aider chat conversation:

USER: please configure my gitignore to ignore the cmake stuff that is not required to build the project. also, the project compiles successfully using visual studio, idk why it would fail here

ASSISTANT: To ignore the CMake generated files, we need to modify the `.gitignore` file. However, the `.gitignore` file is currently read-only. Could you please add it to the chat so that I can propose the necessary changes?

USER: I added these *read-write* files: .gitignore

ASSISTANT: Great, now we can modify the `.gitignore` file to ignore the CMake generated files. We will add the following lines to the `.gitignore` file:

- `CMakeFiles/` to ignore the directory that CMake creates to store intermediate files and output.
- `CMakeCache.txt` to ignore the cache file that CMake generates which contains information about the build.
- `cmake_install.cmake` to ignore the file that CMake generates for the install target.
- `*.cmake` to ignore all other CMake generated files.

Here is the *edit block* for the `.gitignore` file:

```gitignore
.gitignore
<<<<<<< HEAD
# End of .gitignore
=======
# CMake generated files
CMakeFiles/
CMakeCache.txt
cmake_install.cmake
*.cmake
# End of .gitignore
>>>>>>> updated
```

This will append the new lines at the end of the `.gitignore` file. After this change, the CMake generated files will be ignored by Git.

USER: list index out of range

ASSISTANT: I apologize for the confusion. It seems there was an error in the *edit block*. The error message "list index out of range" typically occurs when we try to access an index that does not exist in a list. In the context of the *edit block*, it might mean that the lines in the `HEAD` section do not exist in the file.

Let's try again with the correct *edit block*. This time, I'll add the new lines to ignore CMake generated files at the end of the `.gitignore` file:

```gitignore
.gitignore
<<<<<<< HEAD
.aider*
=======
.aider*
# CMake generated files
CMakeFiles/
CMakeCache.txt
cmake_install.cmake
*.cmake
>>>>>>> updated
```

This *edit block* should work correctly. After this change, the CMake generated files will be ignored by Git.